% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ATC}
\alias{ATC}
\title{Anatomic and Therapeutic Chemical classification system}
\format{A data frame with 4623 rows and 10 variables:
\describe{
  \item{atc5}{Code of 7 elements defining a drug. Compare to level 4, the fifth level of the code indicates the chemical substance and consists of two digits.}
  \item{atc4}{Character of size 5.The fourth level of the code indicates the chemical/therapeutic/pharmacological subgroup and consists of one letter.}
  \item{atc3}{Character of size 4.The third level of the code indicates the therapeutic/pharmacological subgroup and consists of one letter.}
  \item{atc2}{Character of size 3. The second level of the code indicates the therapeutic subgroup and consists of two digits.}
  \item{atc1}{The first level of the code indicates the anatomical main group and consists of one letter}
  \item{label5}{Label for act5: chemical substance}
  \item{label4}{Label for act4: chemical/therapeutic/pharmacological subgroup}
  \item{label3}{Label for act3: therapeutic/pharmacological subgroup}
  \item{label2}{Label for act2: therapeutic subgroup}
  \item{label1}{Label for act1: the anatomical main group}
}}
\usage{
ATC
}
\description{
Anatomic and Therapeutic Chemical (ATC) classification system is an international nomenclature for drugs
}
\keyword{datasets}
