% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio2rdf_db.R
\name{bio2rdf_db}
\alias{bio2rdf_db}
\title{Retrieve drug informations from bio2rdf}
\usage{
bio2rdf_db(lang = "en", atc = NULL, db = NULL)
}
\arguments{
\item{lang}{Character vector to specify the language of the results returned}

\item{atc}{A vector of atc codes for which to retrieve informations}

\item{db}{A vector of db codes for which to retrieve informations}
}
\description{
Retrieve drug informations from bio2rdf SPARQL endpoint such as label, description, category, etc.
}
\examples{
 \dontrun{
   bio2rdf <- bio2rdf_db(lang="en")
   head(bio2rdf)
}

}
\author{
Y. Rivault
}
