% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes2cui.R
\name{codes2cui}
\alias{codes2cui}
\title{Mapping medical nomenclatures to CUI}
\usage{
codes2cui(codes = NULL, ontologies = "", api_key = "")
}
\arguments{
\item{codes}{Codes the user want to map. It can be cui's codes or any terminology codes available on Bioportal.}

\item{ontologies}{A set of ontologies where to find mappings if codes are cui. Or a set of ontologies which codes belongs if they are not cui.}

\item{api_key}{An API Key is required to access any API call on BioPortal and SIFR. To have one, you need to register at \href{http://bioportal.bioontology.org/}{BioPortal}, respectively to \href{http://bioportal.lirmm.fr/}{SIFR BioPortal}.}
}
\value{
data.frame of mapped codes
}
\description{
The function allows mapping from concept unique identifier (cui - UMLS) to medical nomenclatures available through \href{http://bioportal.bioontology.org/}{BioPortal}, and vice versa.
}
\examples{
\dontrun{
 mcui <-  codes2cui(codes="B01AA01",ontologies="ATC",api_key="youAPIKey")
 mcui
 }
}
\author{
Y. Rivault
}
