% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disease_set}
\alias{disease_set}
\title{Disease set}
\format{dataframe  1997 rows by 4 variables that contains patients Id and prescribed drugs, codified according to the ICD10 and mapped to cui and NDFRT.}
\usage{
disease_set
}
\description{
Example  of patients diagnosis
}
\author{
Y. Rivault
}
\keyword{datasets}
