% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_set}
\alias{drug_set}
\title{Drug set}
\format{dataframe  13421 rows by 4 variables that contains patients Id and prescribed drugs, codified according to the ATC and mapped to cui and NDFRT.}
\usage{
drug_set
}
\description{
Example of prescribed drug to patients
}
\author{
Y. Rivault
}
\keyword{datasets}
