\name{find_pairs}
\alias{find_pairs}

\title{
Find pairs of relation
}
\description{
Internal function for the find_relations function call
}
\usage{
find_pairs(items = "", target = "", target_elements = "")
}
\arguments{
 \item{items}{
  database where to search for relations.
  }
  \item{target}{
  Knowledge database (e.g. DIKB, DID, NDF-RT, etc)
  }
  \item{target_elements}{
  Rows representing the relations.
  }
}
\details{
No meant to be used by itself
}
\author{
Y. Rivault
}
\note{
internal function
}
\seealso{
See Also as \code{\link[pkg:queryMed]{find_relations}}
}

