% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_relations.R
\name{find_relations}
\alias{find_relations}
\title{Find relation between medical terms}
\usage{
find_relations(data.x, data_indices, data.y = NULL,
  data_elements.x = NULL, data_elements.y = NULL, target,
  target_elements, progress = "none")
}
\arguments{
\item{data.x}{Data frame with one column of identifier and one column of medical terms}

\item{data_indices}{column names of data.x that contains identifiers.}

\item{data.y}{Data frame with one column of identifier and one column of medical terms}

\item{data_elements.x}{column names of data.x that contains medical term of interest.}

\item{data_elements.y}{column names of data.y that contains medical term of interest.}

\item{target}{Knowledge database (e.g. DIKB, DID, NDF-RT, etc)}

\item{target_elements}{Column names where to look for relations (in pairs).}

\item{progress}{Whether or not to display progress bar. Default is set to "none"}
}
\description{
Function to find intersection between medical terms (disease, drugs...) and pairs defining a biomedical relationship like interactions, contraindications and indications listed in Knowledge database.
}
\examples{
# Do patients have drug interaction
data(drug_set)
data(DIKB)
interactions <- find_relations(data.x=drug_set,   data_indices = "patient",  
data_elements.x = "drugs", target=DIKB, target_elements = c("atc1","atc2"),  progress="none")
 interactions[[1]]
}
\references{
Ayvaz et al (2015) Toward complete dataset of drug-drug interaction information from publicly available sources. Journal of Biomedical Informatics, 55: 206-217.

Sharp ME (2017). Toward a comprehensive drug ontology: extraction of drug-indication relations from diverse information sources. Journal of Biomedical Semantics,8:2.
}
\seealso{
[DIKB] dataset and [DID] dataset
}
\author{
Y. Rivault
}
