% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DIKB.R
\name{get_DIKB}
\alias{get_DIKB}
\title{Retrieve Drug Interaction Knowledge Base (DIKB) data}
\usage{
get_DIKB(path = getwd(), url = NULL, mapping = NULL)
}
\arguments{
\item{path}{filepath where to download the Drug Interaction Knowledge Base. By default it is the current working directory.}

\item{url}{URL of the location of the Drug Interaction Knowledge Base (DIKB). By default it is <https://dbmi-icode-01.dbmi.pitt.edu/dikb-evidence/pddi-sets/> but the argument exists in case the url happens to change.}

\item{mapping}{Optional. Possible argument is ATC and the mapping is performed between DIKB and ATC information from  <http://bio2rdf.org and https://dbpedia.org/>}
}
\value{
Data frame with all variables available from Drug Interaction Knowledge Base (DIKB).
}
\description{
This function downloads the Drug Interaction Knowledge Base (DIKB) from source.
}
\details{
The function download the less conservative version of all PDDI datasets merged by Richard D. Boyce's group at University of Pittsburgh. See <https://dbmi-icode-01.dbmi.pitt.edu/dikb-evidence/pddi-sets/> for details.
}
\examples{
\dontrun{
DIKB <- get_DIKB(path="/tmp",mapping="ATC")
head(DIKB)
colnames(DIKB)
dim(DIKB)
}
}
\references{
Ayvaz et al (2015) Toward complete dataset of drug-drug interaction information from publicly available sources. Journal of Biomedical Informatics, 55: 206-217.
}
\seealso{
[get_DIKB()] and [mapping_atc_db()]
}
\author{
Y. Rivault
}
