% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{is.atc}
\alias{is.atc}
\alias{is.icd10}
\title{Test validity of ATC or ICD10 code}
\usage{
is.atc(x)

is.icd10(x)
}
\arguments{
\item{x}{Character vector to test as ATC code.}
}
\value{
logical vector that indicates which elements are ATC code
}
\description{
`is.atc` verify that the given character vector is an ATC code
}
\examples{
is.atc(c("C10AB01", "N07BB01"))
is.atc(c("C10AB01", "N07"))
}
