% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_atc_db.R
\name{mapping_atc_db}
\alias{mapping_atc_db}
\title{Mapping the Anatomical Therapeutic and Chemical classidication to DrugBank}
\usage{
mapping_atc_db()
}
\value{
Return a data.table of size 2993 rows by 2  columns giving the whole Anatomical Therapeutic and Chemical classification to DrugBank mapping.
}
\description{
The function allows to get all the mapping from the Anatomical Therapeutic and Chemical classification to DrugBank, through the bio2rdf and dbpedia sparql endpoints.
}
\details{
No argument needed.
}
\examples{
  \dontrun{
    xx = mapping_atc_db()
    head(xx)
  }
}
\references{
Alison Callahan, Jose Cruz-Toledo, Peter Ansell, Michel Dumontier: Bio2RDF Release 2: Improved Coverage, Interoperability and Provenance of Life Science Linked Data. ESWC 2013: 200-212.

Lehman, J et al (2015) Dbpedia: a large-scale, multilingual knowledge extracted from wikipedia. Semantic Web, 6: 167-195
}
\seealso{
[get_DID]
}
\author{
Y. Rivault
}
