% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pddi_heatmap.r
\name{pddi_heatmap}
\alias{pddi_heatmap}
\title{drug-drug interaction visualization
  
 This function proposes a heatmap representation of drug-drug interaction as listed in DIKB data base}
\usage{
pddi_heatmap(data, object, level.precipitant)
}
\arguments{
\item{data}{character object that specifies data base name}

\item{object}{character vector that specifies drug name(s)}

\item{level.precipitant}{numeric that states the interaction level number between choosen object and precipitants (between 1 and 5, from largest ATC level to precipitant drugs name)}
}
\value{
A heatmap plot is returned.
}
\description{
drug-drug interaction visualization
  
 This function proposes a heatmap representation of drug-drug interaction as listed in DIKB data base
}
\examples{
data(DIKB)
pddi_heatmap(data = DIKB, object = "CLOPIDOGREL", level.precipitant = 4)
pddi_heatmap(data = DIKB, object = c("TADALAFIL","MORPHINE"), level.precipitant = 1)
}
\seealso{
[DIKB] dataset
}
\author{
S. Tessier
}
