\name{queryMed-package}
\alias{queryMed-package}
\alias{queryMed}
\docType{package}
\title{
\packageTitle{queryMed}
}
\description{
\packageDescription{queryMed}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{queryMed}
\packageIndices{queryMed}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{queryMed}

Maintainer: \packageMaintainer{queryMed}
}
\references{
~~ Literature or other references for background information ~~
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
