% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search}
\alias{search}
\title{Access to BioPortal or SIFR search REST API}
\usage{
search(term = "", ontologies = "", service = "bioportal",
  api_key = "", extra_args = "")
}
\arguments{
\item{term}{A list of terms to search for informations on Bioportal. It can be codes from nomenclatures.}

\item{ontologies}{A list of ontologies to constrain the search of informations.}

\item{service}{Character vector that specifies the web service to query either "bioportal" or "sifr".}

\item{api_key}{An API Key is required to access any API call on BioPortal and SIFR. To have one, you need to register at <http://bioportal.bioontology.org/>, respectively to <http://bioportal.lirmm.fr/> for SIFR.}

\item{extra_args}{Extra arguments to parse into the query.}
}
\value{
The return value is a list of lists of elements that define the request term
}
\description{
This function allows retrieving knowledge from medical ontologies through BioPortal or SIFR search REST API.
}
\details{
This is a low level function. The return are return as JSON list of lists.
}
\examples{
 \dontrun{
    disease72 = queryMed::search(term = "I72", ontologies = "ICD10", 
    service = "bioportal", api_key = "yourAPIkey")
    disease72$collection[[1]]$prefLabel
 }   
  
}
\references{
Noy NF, Shah NH, Whetzel PL, Dai B, Dorf M, Griffith N, Jonquet C, Rubin DL, Storey MA, Chute CG, Musen MA. BioPortal: ontologies and integrated data resources at the click of a mouse. Nucleic Acids Res. 2009 Jul;37 (Web Server issue):W170-3.

Whetzel PL, Noy NF, Shah NH, Alexander PR, Nyulas C, Tudorache T, Musen MA. BioPortal: enhanced functionality via new Web services from the National Center for Biomedical Ontology to access and use ontologies in software applications. Nucleic Acids Res. 2011 Jul;39(Web Server issue):W541-5. Epub 2011 Jun 14.

Salvadores M, Horridge M, Alexander PR, Fergerson RW, Musen MA, and Noy NF. Using SPARQL to Query BioPortal Ontologies and Metadata. International Semantic Web Conference. Boston US. LNCS 7650, pp. 180195, 2012.
}
\author{
Y. Rivault
}
