% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbpedia_drug.R
\name{dbpedia_drug}
\alias{dbpedia_drug}
\title{Information retrieval from DBpedia drug}
\usage{
dbpedia_drug(lang = "en", atc = NULL, db = NULL)
}
\arguments{
\item{lang}{Language for character results}

\item{atc}{A vector of ATC codes for which to retrieve informations}

\item{db}{A vector of DrugBank codes for which to retrieve informations}
}
\value{
data table of drug terms and their annotation : drug, atc code,  drug bank identifier,  abstract,   comment, label. When querying the whole dataset, the data table is of size 8712 rows by 6 columns.
}
\description{
Retrieve drug information from dbpedia such as atc and drug bank identifier codes
}
\examples{
  \dontrun{
    drug_dbp <- dbpedia_drug(lang="en")
    head(drug_dbp)
  }
}
\references{
Lehman, J et al (2015) Dbpedia: a large-scale, multilingual knowledge extracted from wikipedia. Semantic Web, 6: 167-195
}
\seealso{
[dbpedia_disease]
}
\author{
Y. Rivault
}
