% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DID}
\alias{DID}
\title{Drug Indication Database
 
 Curated version of the Drug Indication Database (DID) downloaded from supplementary material of Sharp ME (2017).}
\format{Data frame of 66 variables by 464 056 rows.
 \describe{ 
 Drugs and their phenotype indications (first two variables) are coded according to Concept Unique Identifier (CUI) codification from the Unified Medical Language System (UMLS). The other variables define their characteristics.
 }}
\usage{
DID
}
\description{
Drug Indication Database
 
 Curated version of the Drug Indication Database (DID) downloaded from supplementary material of Sharp ME (2017).
}
\references{
harp ME (2017). Toward a comprehensive drug ontology: extraction of drug-indication relations from diverse information sources. Journal of Biomedical Semantics,8:2.

Sharp ME (2017) supplementary material: <https://figshare.com/articles/Additional_file_1_of_Toward_a_comprehensive_drug_ontology_extraction_of_drug-indication_relations_from_diverse_information_sources/4535021>
}
\author{
Y. Rivault and N. Le Meur
}
\keyword{datasets}
