% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DIKB}
\alias{DIKB}
\title{Drug Interaction Knowledge Base data
 
 Curated version of the Drug Interaction Knowledge Base (DIKB). The modalities of the various variables depend on the  data source (ie. KEGG, NDFRT, ANSM). See reference for more details}
\format{Data frame of 15 variables by 202 363 rows.
 \describe{
   \item{drug2}{Drug bank ID of the second (precipitant) interacting drug}
   \item{drug2}{Drug bank ID of the first (object) interacting drug}
   \item{object}{Full name of the the first interacting drug}
   \item{precipitant}{Full name of the the second interacting drug}
   \item{contraindication}{Logical vector that states whether or not it is a conterindication according to the KEGG database only}
   \item{precaution}{Logical vector that states whether or not precaution are needed according to the KEGG database only}
   \item{severity}{Severity of the interactions coded differently according to the different sources. See references for details}
   \item{uri}{Source uri}
   \item{source}{Database from which are issued the data}
   \item{evidenceStatement}{Raw text that states the evidence of the interaction}
   \item{atc1}{ATC code the first (object) interacting drug}
   \item{atc2}{ATC code the second (precipitant) interacting drug}
 }}
\usage{
DIKB
}
\description{
Drug Interaction Knowledge Base data
 
 Curated version of the Drug Interaction Knowledge Base (DIKB). The modalities of the various variables depend on the  data source (ie. KEGG, NDFRT, ANSM). See reference for more details
}
\references{
Ayvaz et al (2015) Toward complete dataset of drug-drug interaction information from publicly available sources. Journal of Biomedical Informatics, 55: 206-217.

Richard D. Boyces group at University of Pittsburgh. See <https://dbmi-icode-01.dbmi.pitt.edu/dikb-evidence/pddi-sets/>
}
\author{
Y. Rivault and N. Le Meur
}
\keyword{datasets}
