% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DIKB}
\alias{DIKB}
\title{Drug Interaction Knowledge Base data
 
 Curated version of the Drug Interaction Knowledge Base (DIKB). The modalities of the various variables depend on the  data source (ie. KEGG, NDFRT, ANSM). See reference for more details}
\format{Data frame of 26 variables by 91 553 rows.
 \describe{
   \item{drug.pairs}{Unique drug bank ID pair (object*precipitant)}
   \item{Kegg}{Kegg precaution evaluation of a drug bank ID pair (between 1 and 3)}
   \item{FrenchDB}{FrenchDB precaution evaluation of a drug bank ID pair (between 1 and 3)}
   \item{NDFRT}{NDFRT precaution evaluation of a drug bank ID pair (between 1 and 3)}
   \item{OSCAR}{OSCAR precaution evaluation of a drug bank ID pair (between 1 and 3)}
   \item{WorldVista}{WorldVista precaution evaluation of a drug bank ID pair (between 1 and 3)}
   \item{mean}{Mean of the 5 precaution evaluations of the 5 different sources (between 1 and 3)}
   \item{max}{Maximum of the 5 precaution evaluations of the 5 different sources (between 1 and 3)}
   \item{diff.mean.max}{Difference between mean and max of the 5 precaution evaluations of the 5 different sources (between 0 and 1.5)}
   \item{score.max}{Final score of precaution for a pair of drug bank ID (calcul : if(mean>4/3){max-1/10*diff.mean.max)} else{max-diff.mean.max}}
   \item{drug.id.object}{Drug bank ID of the first (object) interacting drug}
   \item{drug.id.precipitant}{Drug bank ID of the second (precipitant) interacting drug}
   \item{atc.object}{ATC code the first (object) interacting drug}
   \item{atc.precipitant}{ATC code the second (precipitant) interacting drug}
   \item{label1.atc.object}{Anatomical main group of the object ATC code}
   \item{label1.atc.object}{Anatomical main group of the object ATC code}
   \item{object}{Full name of the the first interacting drug}
   \item{precipitant}{Full name of the the second interacting drug}
   \item{label2.atc.object}{therapeutic subgroup of the object ATC code}
   \item{label2.atc.precipitant}{therapeutic subgroup of the object ATC code}
   \item{label3.atc.object}{therapeutic/pharmacological subgroup of the object ATC code}
   \item{label3.atc.precipitant}{therapeutic/pharmacological subgroup of the object ATC code}
   \item{label4.atc.object}{chemical/therapeutic/pharmacological subgroup of the object ATC code}
   \item{label4.atc.precipitant}{chemical/therapeutic/pharmacological subgroup of the object ATC code}
   \item{label5.atc.object}{chemical substance of the object ATC code}
   \item{label5.atc.precipitant}{chemical substance of the object ATC code}
 }}
\usage{
DIKB
}
\description{
Drug Interaction Knowledge Base data
 
 Curated version of the Drug Interaction Knowledge Base (DIKB). The modalities of the various variables depend on the  data source (ie. KEGG, NDFRT, ANSM). See reference for more details
}
\references{
Ayvaz et al (2015) Toward complete dataset of drug-drug interaction information from publicly available sources. Journal of Biomedical Informatics, 55: 206-217.

Richard D. Boyces group at University of Pittsburgh. See <https://dbmi-icode-01.dbmi.pitt.edu/dikb-evidence/pddi-sets/>
}
\author{
Y. Rivault, N. Le Meur and S. Tessier
}
\keyword{datasets}
