% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_pairs.R
\name{find_pairs}
\alias{find_pairs}
\title{Find pairs of relation}
\usage{
find_pairs(items, target, target_elements)
}
\arguments{
\item{items}{database where to search for relations.}

\item{target}{Knowledge database (e.g. DIKB, DID, NDF-RT, etc)}

\item{target_elements}{Rows representing the relations.}
}
\description{
Internal function for the find_relations function call
}
\details{
No meant to be used by itself
}
\note{
internal function
}
\seealso{
[find_relations()]
}
\author{
Y. Rivault
}
