\name{find_pairs}
\alias{find_pairs}

\title{
Find pairs of relation
}
\description{
Internal function for the find_relations function call
}
\usage{
find_pairs(items = "", target = "", target_elements = "")
}
\arguments{
 \item{items}{
  database where to search for relations.
  }
  \item{target}{
  Knowledge database (e.g. DIKB, DID, NDF-RT, etc)
  }
  \item{target_elements}{
  Rows representing the relations.
  }
}
\details{
No meant to be used by itself
}
\value{
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Y. Rivault
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
