% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ancestors.R
\name{get_ancestors}
\alias{get_ancestors}
\title{Hierachical search for ancestors from ontologies}
\usage{
get_ancestors(term = "", ontology = "", api_key = "")
}
\arguments{
\item{term}{Term as character vector}

\item{ontology}{Name of the ontology as character vector}

\item{api_key}{An API Key is required to access any API call on Bioportal and SIFR. To have one, you need to register at http://bioportal.bioontology.org/, respectively to http://bioportal.lirmm.fr/ for SIFR.}
}
\value{
Returns a list of ancestors with for instance id as URI with ancestors' code and prefLabel as Labels of the retrieve term
}
\description{
Retrieving a list of ancestors of a class from ontologies available on Bioportal or SIFR web service.
}
\details{
The available ontologies are for instance: ATC, ICD10, NDFRT,SNOMEDCT. See http://bioportal.bioontology.org/ for more details.
}
\examples{
  \dontrun{
    get_ancestors("A05AA",ontology="ATC",api_key = "yourAPIkey")
    get_ancestors("C50",ontology="ICD10",api_key = "yourAPIkey")
    get_ancestors("N0000153235",ontology="NDFRT",api_key = "yourAPIkey")
  }
}
\references{
Noy NF, Shah NH, Whetzel PL, Dai B, Dorf M, Griffith N, Jonquet C, Rubin DL, Storey MA, Chute CG, Musen MA. BioPortal: ontologies and integrated data resources at the click of a mouse. Nucleic Acids Res. 2009 Jul;37 (Web Server issue):W170-3.

Whetzel PL, Noy NF, Shah NH, Alexander PR, Nyulas C, Tudorache T, Musen MA. BioPortal: enhanced functionality via new Web services from the National Center for Biomedical Ontology to access and use ontologies in software applications. Nucleic Acids Res. 2011 Jul;39(Web Server issue):W541-5. Epub 2011 Jun 14.
}
\author{
Y. Rivault
}
