% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DID.R
\name{get_DID}
\alias{get_DID}
\title{Retrieve Drug Indication Database (DID) data}
\usage{
get_DID(mapping_atc = FALSE, mapping_icd10 = FALSE, api_key = "")
}
\arguments{
\item{mapping_atc}{By default the argument is FALSE. When TRUE an additional mapping to ATC code is performed but a valid API Key is required to connect to and query BIOPORTAL services.}

\item{mapping_icd10}{By default the argument is FALSE. When TRUE an additional mapping to ICD10 code is performed but a valid API Key is required to connect to and query BIOPORTAL services.}

\item{api_key}{Optional argument. A valid API Key is required to access any API call on Bioportal. To have one, you need to register at http://bioportal.bioontology.org/.}
}
\value{
Data frame with all Drug Indication Database variables and possible mapping to ATC and/or ICD10 if arguments are set to TRUE and a valid API Key is provided.
}
\description{
Function that load Drug Indication Database (DID) from source.
}
\details{
See https://figshare.com/articles/Additional_file_1_of_Toward_a_comprehensive_drug_ontology_extraction_of_drug-indication_relations_from_diverse_information_sources/4535021 for details
}
\examples{
  DID <- get_DID(mapping_atc=FALSE,mapping_icd10=FALSE)
  head(DID)
  dim(DID)
  colnames(DID)

  # If you want only indications from ATC to ICD10 codes :
  atc_icd10_DID <- DID[!(is.na(DID$atc) | is.na(DID$icd10)),]
}
\references{
Sharp ME (2017). Toward a comprehensive drug ontology: extraction of drug-indication relations from diverse information sources. Journal of Biomedical Semantics,8:2.
}
\seealso{
[get_DIKB()] and [mapping_cui()]
}
\author{
Y. Rivault
}
